rename acurl tests
