+++ acurl/tests/test_httpbin.py
@@ -14,7 +14,7 @@ async def session():
 @pytest.mark.asyncio
 async def test_get(httpbin):
     s = await session()
-    r = await s.get(httpbin.url + "/ip")
+    r = await s.get(f"{httpbin.url}/ip")
     assert r.status_code == 200
     assert isinstance(r.headers, dict)
     # FIXME: is this a method in the requests api?
@@ -23,13 +23,13 @@ async def test_get(httpbin):
 
 @pytest.mark.asyncio
 async def test_cookies(httpbin, acurl_session):
-    r = await acurl_session.get(httpbin.url + "/cookies/set?name=value")
+    r = await acurl_session.get(f"{httpbin.url}/cookies/set?name=value")
     assert r.cookies == {"name": "value"}
 
 
 @pytest.mark.asyncio
 async def test_session_cookies(httpbin, acurl_session):
-    await acurl_session.get(httpbin.url + "/cookies/set?name=value")
+    await acurl_session.get(f"{httpbin.url}/cookies/set?name=value")
     cookies = acurl_session.cookies()
     assert cookies == {"name": "value"}
     acurl_session.erase_all_cookies()
@@ -40,8 +40,8 @@ async def test_session_cookies(httpbin, acurl_session):
 @pytest.mark.asyncio
 async def test_session_cookies_sent_on_subsequent_request(httpbin):
     s = await session()
-    await s.get(httpbin.url + "/cookies/set?name=value")
-    resp = await s.get(httpbin.url + "/cookies")
+    await s.get(f"{httpbin.url}/cookies/set?name=value")
+    resp = await s.get(f"{httpbin.url}/cookies")
     data = resp.json()
     assert len(data) == 1
     assert data["cookies"] == {"name": "value"}
@@ -50,15 +50,19 @@ async def test_session_cookies_sent_on_subsequent_request(httpbin):
 @pytest.mark.asyncio
 async def test_set_cookies(httpbin):
     s = await session()
-    await s.get(httpbin.url + "/cookies/set?name=value")
-    r = await s.get(httpbin.url + "/cookies/set?name2=value", cookies={"name3": "value"})
+    await s.get(f"{httpbin.url}/cookies/set?name=value")
+    r = await s.get(
+        f"{httpbin.url}/cookies/set?name2=value", cookies={"name3": "value"}
+    )
     assert r.cookies == {"name": "value", "name2": "value", "name3": "value"}
 
 
 @pytest.mark.asyncio
 async def test_basic_auth(httpbin):
     s = await session()
-    r = await s.get(httpbin.url + "/basic-auth/user/password", auth=("user", "password"))
+    r = await s.get(
+        f"{httpbin.url}/basic-auth/user/password", auth=("user", "password")
+    )
     assert r.status_code == 200
 
 
@@ -66,7 +70,7 @@ async def test_basic_auth(httpbin):
 async def test_failed_basic_auth(httpbin):
     s = await session()
     r = await s.get(
-        httpbin.url + "/basic-auth/user/password", auth=("notuser", "notpassword")
+        f"{httpbin.url}/basic-auth/user/password", auth=("notuser", "notpassword")
     )
     assert r.status_code == 401
 
@@ -74,6 +78,6 @@ async def test_failed_basic_auth(httpbin):
 @pytest.mark.asyncio
 async def test_redirect(httpbin):
     s = await session()
-    url = httpbin.url + "/ip"
-    r = await s.get(httpbin.url + "/redirect-to?" + urlencode({"url": url}))
+    url = f"{httpbin.url}/ip"
+    r = await s.get(f"{httpbin.url}/redirect-to?" + urlencode({"url": url}))
     assert r.url == url
