+++ acurl/tests/test_request.py
@@ -1,7 +1,7 @@
 from unittest.mock import Mock
 
 import pytest
-from helpers_ng import create_request
+from helpers import create_request
 
 import acurl
 
@@ -36,7 +36,7 @@ def test_request_cookies():
 @pytest.mark.asyncio
 async def test_request_e2e(httpbin, acurl_session):
     r = await acurl_session.get(
-        httpbin.url + "/get", headers={"Foo": "bar"}, cookies={"baz": "quux"}
+        f"{httpbin.url}/get", headers={"Foo": "bar"}, cookies={"baz": "quux"}
     )
     assert r.request.cookies == {"baz": "quux"}
     assert r.request.headers == {"Foo": "bar"}
@@ -44,8 +44,8 @@ async def test_request_e2e(httpbin, acurl_session):
 
 @pytest.mark.asyncio
 async def test_request_cookies_from_previous(httpbin, acurl_session):
-    await acurl_session.get(httpbin.url + "/cookies/set?name=value")
-    r = await acurl_session.get(httpbin.url + "/get", cookies={"foo": "bar"})
+    await acurl_session.get(f"{httpbin.url}/cookies/set?name=value")
+    r = await acurl_session.get(f"{httpbin.url}/get", cookies={"foo": "bar"})
     assert r.request.cookies == {"name": "value", "foo": "bar"}
 
 
@@ -54,10 +54,10 @@ async def test_request_cookies_from_previous(httpbin, acurl_session):
 async def test_request_cookies_from_previous_excludes_other_domains(
     httpbin, acurl_session
 ):
-    await acurl_session.get(httpbin.url + "/cookies/set?name=value")
+    await acurl_session.get(f"{httpbin.url}/cookies/set?name=value")
     # FIXME: we want to set a cookie for another domain.  There's no easy way
     # to get another domain set up loally, so we (as an exception) go out to
     # the network for this test.
     await acurl_session.get("https://httpbin.org/cookies/set?foo=bar")
-    r = await acurl_session.get(httpbin.url + "/get")
+    r = await acurl_session.get(f"{httpbin.url}/get")
     assert r.request.cookies == {"name": "value"}
